requirements.txt
pandas==2.2.2
python-dotenv==0.21.0
requests==2.32.3
streamlit==1.37.1